# services/billing/openapi.yaml
openapi: 3.0.3
info:
  title: Billing Service
  version: 0.1.0
  license:
    name: Elastic-Internal
servers:
  - url: /api/v1/billing
paths:
  /tokens/balance:
    get:
      operationId: getTokenBalance
      summary: Get current user's token balance
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance: { type: integer }
                required: [balance]
        '401': { description: Unauthorized }
...
  /subscription:
    get:
      operationId: getSubscription
      summary: Get current subscription
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '401': { description: Unauthorized }
        '404': { description: Not Found }
  /tokens/reserve:
    post:
      operationId: reserveTokens
      summary: Reserve tokens for a task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: integer }
                taskId: { type: string }
              required: [amount]
      responses:
        '202': { description: Reserved }
        '400': { description: Invalid input }
        '401': { description: Unauthorized }
...
  /tokens/credit/subscription:
    post:
      operationId: creditSubscriptionTokens
      summary: Credit subscription tokens (pool-aware, idempotent)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: integer, minimum: 1 }
                description: { type: string }
                metadata:
                  type: object
                  additionalProperties: true
                  properties:
                    expiresAt: { type: string, format: date-time }
              required: [amount]
      responses:
        '200': { description: Credited }
        '401': { description: Unauthorized }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Subscription:
      type: object
      properties:
        id: { type: string }
        planName: { type: string }
        status: { type: string }
        currentPeriodEnd: { type: string, format: date-time }
      required: [id, planName, status, currentPeriodEnd]
...